pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tcnart_py"
version = "0.1.0"
description = "Python reimplementation of parts of tcnart (core, schema, serialization, network) using zenoh and pycdr2"
authors = [{name = "Artekmed", email = "dev@artekmed"}]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "eclipse-zenoh>=1.3.4",  # Eclipse Zenoh Python (sync API used)
  "pycdr2>=1.0.0",
  "dataclasses-json>=0.6.7; python_version<'3.11'",
  "typing_extensions>=4.10.0",
]

[project.urls]
Homepage = "https://github.com/artekmed/artekmed_rust_client"
